+++ src/deluge/playback/mode/session.h
@@ -34,7 +34,6 @@ public:
 	void doLaunch(bool isFillLaunch);
 	void scheduleLaunchTiming(int64_t atTickCount, int32_t numRepeatsUntil, int32_t armedLaunchLengthForOneRepeat);
 	int32_t getNumSixteenthNotesRemainingTilLaunch();
-	int32_t getNumBarsRemainingTilLaunch();
 	void scheduleFillEvent(Clip* clip, int64_t atTickCount);
 	void cancelAllLaunchScheduling();
 	void launchSchedulingMightNeedCancelling();